fix tryattach; fix mouseretab; refactor code
That else case in tryattach could never been true; it should be
a separated if case.

The mouseretab() function called the function containerredecorate()
instead of containerdecorate().  That means that the decoration was
only redraw when the container was resized.  But the container is never
resized during a mouseretab.  Instead, we need to redraw when a row or
column was deleted.  So I put a "redraw" variable that I set to 1 (true)
when a row or column is deleted, then redecorate the container when this
variable is 1.  I also removed the call to rowstack(), so the user can
retab into a restacked column.

I removed commented out code.  The calls to an outlinedraw() function
were there from a time when I was drawing a outline/frame of the
container while resizing it with the mouse instead of resizing the
container itself.  For more information on how I changed it, see [1].

[1]: https://nixers.net/Thread-window-manager-development-thread#pid23137