add Plan 9's rio hidden windows behavior
On Plan 9's rio, when a window is obscured by other windows, it is
listed as hidden.

I incorpored this behavior in shod.  This behavior however can be
disabled with -h (which is not a help flag).

I also merged the four container lists based on stacking order (one for
each Z-axis layer) into a single container list with dummy containers to
indicate the top of each Z-axis layer.