add automatic tabbing; dettach menu from tab; etc
I'm working on a fork of suckless's surf(1) web browser that uses
shod(1), rather than tabbed(1), for tabbing.  Previously, shod did
tabbing manually: the user needed to drag and drop a window into another
in order for both to be tabbed.  With this commit, a window being mapped
sharing the same leader as another, already mapped, window (or if the
second is the leader of the first), the window being mapped is placed
automatically as a tab next to the other.  Thus, middle clicking a link
in my fork of surf opens the new surf instance in a new tab.

Previously, menus were relative to a single window and could not be
shared.  This behavior breaked gimp's multi window mode (where the
palletes menus are shared between open images).  So if the user has two
images open, the menus only appeared on the window of the first image
(which the menus are relative to), and when the user focus the window of
the next image, the menus disappeared.  With this commit, menus can be
shared between windows.

I also fixed a bug on splash screens that made shod not map new windows
while the splash screen was visible.  Thankfully, splash screens are
short-lived windows, so that bug was only active during the short splash
screen's lifetime.