select SubstructureRedirectMask'd events only once
With the addition of checkotherwm(), we are selecting the events
associated with the mask SubstructureRedirectMask twice: once at
checkotherwm() and another time at initroot().

At initroot(), however, we were doing more than just selecting
SubstructureRedirectMask events: we also select other events
and set the default cursor for the root window.  All in a single
XChangeWindowAttributes(3) call.

With this commit, the XChangeWindowAttributes(3) call is split into
a call to XDefineCursor(3) at initroot() to set the cursor, and a
XSelectInput(3) at checkotherwm() to select all events (including
the ones associated with SubstructureRedirectMask).