deprecate -m; add -A and -W
The -m was problematic.

Shod needs both a key symbol and a modifier to work; and also needs that
such modifier be generated by such key symbol.  Both things were set by
the -m command-line option.  Two things could happen:

- The user provides an invalid key symbol with -m.  Shod just failed on
  such case.

- The user provides a valid key symbol which produces no modifier key
  (say, "space", or "F8").  Shod initiates successfully, but acts
  erradically.

There are only a few modifiers on X:

- Control (should not be used by shod; applications depend on it).
- Shift (should not be used by shod; uppercase letters depend on it).
- Mod1Mask/Alt (shod can use it).
- Mod2Mask/NumLock (should not be used by shod; the keypad depends on it).
- Mod3Mask/ScrllLock (nobody knows what it is or does).
- Mod4Mask/Super/Win (shod can use it).
- Mod5Mask/AltGR (should not be used by shod; typing extra characters on
  some keyboard layouts depends on it).

As you can see, the only modifier keys one can freely use on a window
manager are Mod1Mask/Alt and Mod4Mask/Super.  Therefore, it is not a
problem to limit the options to only allow the user to use one of those
two keys.

The option -m, which was too powerful and only useful when invoked as
`-m Alt_L` or `-m Super_L`, has been then replaced by -A and -W.