undo incremental update of _NET_CLIENT_LIST etc
On 660f988a0a32 I set the _NET_CLIENT_LIST and related properties
incrementally by successive calls of XChangeProperty(3) with mode
PropModePrepend for each such property.

This commit revert to a single call with mode PropModeReplace for
each property.

The reason is that each call to XChangeProperty(3) (regardless of
the mode) generates a PropertyNotify event (see XPropertyEvent(3)
for information on the event).  Clients like pagers and taskbars,
which watches this event for those properties at the root window,
gets overloaded with notifications when using incremental changes
(compared to a single change, which only notifies once).