rewrite decorations; remove lots of lines
Decoration windows (corners, edges, divisions, etc) are redesigned.

Events on those windows are now relative to those windows themselves;
rather than passed to the parent frame window as they previously were.
We need not guess what is being dragged around based on cursor position:
the dragged decoration window is exactly the one specified in the event.
If a left edge is dragged, then the container is resized to the left;
no need to compute what is under the pointer's coordinates.

Decorations dont have a pressed/pushed-down state anymore (except buttons).
Handling them yields ongoing visual feedback while the mouse button is down.
Dragging the bottom-left corner around resizes container in that direction;
there is no need to indicate that the corner is being manipulated.
Only buttons, which is only activated on mouse release, have pressed state.

Decorations are not redrawn whenever the container state or size changes.
All possible states are drawn on a set of pixmaps, which are then copied
over the decorations when needed.

                                 * * *

It has been more than a year since i touched much of these cold...
It's uglier than i thought.  Gonna throw most of it on the trash.