permit refocusing focused tab on all (but one) contexts
Move test of whether the focused tab is already the focused one from
tabfocus() to tab_btnpress(), as the main issue is for calling
tabfocus() whenever a button press ocurrs on a client window.

Calling tabfocus() on a focused tab is needed in some cases, to
re-set changed attributes of a container.  For example, when attaching
a focused tab into a different container, it keeps being the focused
tab, but the `.seltab` member (the pointer to which tab of a row is the
active one) must be updated, and that is done by tabfocus().